FinalVersionFoodgrams

+++ app/Http/Controllers/PostController.php
@@ -14,6 +14,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 
+
 class PostController extends Controller
 {
     private $post;
@@ -27,12 +28,11 @@ class PostController extends Controller
     }
     public function index()
     {
-        
+   
         return Inertia::render('Dashboard',[
             'posts' => Posts::getPost(Auth::id())
            ]);  
-        
-        //return   Posts::getPost(Auth::id());
+
     }
 
     public function create(){
@@ -41,9 +41,7 @@ class PostController extends Controller
 
     }
 
-    public function getPosts(){
-      
-    }
+    
 
 
 
@@ -58,28 +56,40 @@ class PostController extends Controller
 
     public function show(Posts $post, User $user)
     {
-        $idAuth=Auth::id();
-        if($idAuth == $post->user_id || ['role:Admin'])
-        {
-            return Inertia::render('Post/Edit',[
-                'PostEdit' => $post
-             ]);
-        }
+        
        
-        return redirect() -> route('dashboard');  
+       if($post->user_id !== Auth::id())
+        {
+            return redirect()->route('dashboard');
+        }
+        
+                
+       return Inertia::render('Post/Edit',[
+           'PostEdit' => $post
+       ]);
+      
         
        
     }
 
     public function view(Request $request)
     {
+        
         $id = $request->post;
+
        
         $post = Posts::getPostView($id);
+
+        if(count($post) === 0)
+        {
+            return redirect()->route('dashboard');
+
+        }
         
-        return Inertia::render('Post/View',[
+        return $post[0];
+        /*return Inertia::render('Post/View',[
             'post' => $post[0]
-        ]);
+        ]);*/
        
     }
 
@@ -115,34 +125,29 @@ class PostController extends Controller
 
     public function destroy(Posts $post, User $user)
     {
+        if($post->user_id !== Auth::id())
+        {
+            return redirect()->route('dashboard');
+        }
             
+
             $likes = $post->likes;
-        
+
             foreach($likes as $item) { //foreach element in $arr
                 $deleteLikes = Likes::find($item['id']);
                 $deleteLikes ->delete(); 
- 
             }
             
             $comments = $post->comments;
-            
 
             foreach($comments as $item) { 
                 $deleteComments = Comments::find($item['id']);
                 $deleteComments ->delete(); 
-
             }
             $url = str_replace('storage', 'public', "post");
             Storage::delete( $url,$post->image_path);
             $post->delete(); 
-            return back();
-       
-
-            
-
-     
-        
-               
+            return back();         
     }
 
     public function likeOrDislike(Request $request){
@@ -175,5 +180,84 @@ class PostController extends Controller
             return response()->json($e->getMessage(),500);
         }
     }
+
+    public function showAllPost()
+    {
+        $posts = Posts::all();
+
+        return $posts;
+
+    }
+
+
+    //Post Controller Api
+
+    public function updatePostApi(Request $request)
+    {
+       $editpost = Posts::find($request->id);
+         if($file = $request->file('image')){
+            Storage::delete('public/'. $request->image_path);
+            $file = $request->file('image');
+            $name = uniqid().$file->getClientOriginalName();
+            $storage = Storage::disk('public')->put($name,$file);
+            $request['image_path'] = asset('storage/'.$storage);
+
+        }else{
+            unset($request['image']);
+        }
+
+        $editpost->title = $request->title;
+        $editpost->image_path =$request->image_path;
+        $editpost->description=$request->description;
+        $editpost->ingredients= $request->ingredients;
+        $editpost->portions = $request->portions;
+        $editpost->date_post = Carbon::now();
+
+  
+        $editpost ->save();
+        return with(['msg' => 'Se actualizo el post']);
+
+       
+        
+    }
+
+
+
+
+    
+    public function showPostApi(Posts $post, User $user)
+    {
+        
+         return  $post;
+        
+       
+    }
+
+    public function destroyPostApi(Posts $post, User $user)
+    {
+            $likes = $post->likes;
+
+            foreach($likes as $item) { //foreach element in $arr
+                $deleteLikes = Likes::find($item['id']);
+                $deleteLikes ->delete(); 
+            }
+            
+            $comments = $post->comments;
+
+            foreach($comments as $item) { 
+                $deleteComments = Comments::find($item['id']);
+                $deleteComments ->delete(); 
+            }
+            $url = str_replace('storage', 'public', "post");
+            Storage::delete( $url,$post->image_path);
+            $post->delete(); 
+            return  with(['msg'=>'Se elimino el post']);  
+    }
+
+    public function createApi()
+    {
+        return  with(['msg'=>'Se Ingreso a la vista']);
+    }
+    
     
 }
